Put GroupAlignment in a dedicated alignment folder
